refactor(winscard): restructured the crate
