+++ winscard-rs/src/tlv_tags.rs
@@ -1,0 +1,15 @@
+pub const DATA: u8 = 0x53;
+pub const FASC_N: u8 = 0x30;
+pub const GUID: u8 = 0x34;
+pub const EXPIRATION_DATE: u8 = 0x35;
+pub const ISSUER_SIGNATURE: u8 = 0x3E;
+pub const ERROR_DETECTION_CODE: u8 = 0xFE;
+pub const APPLICATION_PROPERTY_TEMPLATE: u8 = 0x61;
+pub const APPLICATION_IDENTIFIER: u8 = 0x4F;
+pub const COEXISTING_TAG_ALLOCATION_AUTHORITY: u8 = 0x79;
+pub const TAG_LIST: u8 = 0x5C;
+pub const CERTIFICATE: u8 = 0x70;
+pub const CERT_INFO: u8 = 0x71;
+pub const DYNAMIC_AUTHENTICATION_TEMPLATE: u8 = 0x7C;
+pub const DAT_CHALLENGE: u8 = 0x81;
+pub const DAT_RESPONSE: u8 = 0x82;
